move analysis code to analysis.py
